Updates for project layout.
