+++ mTree/microeconomic_system/__init__.py
@@ -1,5 +1,5 @@
 from .agent import Agent
 from .environment import Environment
 from .institution import Institution
-from .dispatcher import Dispatcher
+# from .dispatcher import Dispatcher
 from .property_decorators import *
