+++ mTree/utilities/console/mtree_console_app.py
@@ -1,0 +1,69 @@
+from textual.app import App, ComposeResult
+from textual.screen import Screen
+from textual.containers import Container, Horizontal, VerticalScroll
+from textual.widgets import (
+    Header,
+    Footer,
+    Tab,
+    Tabs,
+    Static,
+    Button,
+    TabbedContent,
+    TabPane,
+    OptionList,
+    Placeholder,
+    ListItem,
+    ListView,
+    Label,
+)
+from textual.widgets.option_list import Option, Separator
+from rich import box
+from rich.console import RenderableType
+from rich.json import JSON
+from rich.markdown import Markdown
+from rich.markup import escape
+from rich.pretty import Pretty
+from rich.syntax import Syntax
+from rich.table import Table
+from rich.text import Text
+
+from mTree.utilities.console.dashboard_screen import DashboardScreen
+from mTree.utilities.console.system_status_screen import SystemStatusScreen
+
+
+
+class MTreeConsoleApp(App):
+    """A Textual app to manage stopwatches."""
+
+    CSS_PATH = "option_list.tcss"
+    BINDINGS = [
+        ("ctrl+d", "quit", "Quit"),
+        ("ctrl+s", "switch_mode('system_status')", "System Status"),
+        ("ctrl+b", "switch_mode('dashboard')", "Dashboard"),
+        ("ctrl+l", "switch_mode('library')", "MES Library"),
+        ("ctrl+h", "switch_mode('help')", "Help"),
+    ]
+
+    MODES = {
+        "dashboard": DashboardScreen,
+        # "library": MTreeRunSimulationScreen,
+        # "help": HelpScreen,
+        "system_status": SystemStatusScreen,
+    }
+
+    def on_mount(self) -> None:
+        self.switch_mode("dashboard")
+
+    # def compose(self) -> ComposeResult:
+    #     """Create child widgets for the app."""
+    #     yield Header()
+    #     with TabbedContent(initial="simulations"):
+    #         with TabPane("Simulations", id="simulations"):  # First tab
+    #             yield Simulations()
+    #         with TabPane("Server Status", id="status"):
+    #             yield Simulations()
+
+    #     yield Footer()
+
+    def action_quit(self) -> None:
+        exit()
