+++ mTree/utilities/console/system_status_screen.py
@@ -4,8 +4,10 @@ import sys
 import pyfiglet
 
 # from mTree.runner.runner import Runner
-from mTree.server.actor_system_startup import ActorSystemStartup
-from mTree.server.actor_system_connector import ActorSystemConnector
+# from mTree.server.actor_system_startup import ActorSystemStartup
+# from mTree.server.actor_system_connector import ActorSystemConnector
+
+from mTree.system.actor_system_connector import ActorSystemConnector
 
 # from thespian.actors import *
 import time
@@ -70,7 +72,7 @@ ROWS = [
 ]
 
 
-from mTree.server.actor_system_connector import ActorSystemConnector
+# from mTree.server.actor_system_connector import ActorSystemConnector
 
 # table_data = [
 #             ['Run Code', 'Configuration', 'Run Number', 'Status', 'Total Time'],
@@ -87,7 +89,7 @@ from mTree.server.actor_system_connector import ActorSystemConnector
 #         print(table.table)
 
 
-class MTreeSystemStatusScreen(Screen):
+class SystemStatusScreen(Screen):
     _refresh_timer: Timer | None
 
     def compose(self) -> ComposeResult:
@@ -98,11 +100,27 @@ class MTreeSystemStatusScreen(Screen):
 
     def on_mount(self) -> None:
         table = self.query_one(DataTable)
-        table.add_columns(
-            "Run Code", "Configuration", "Run Number", "Status", "Total Time"
-        )
+        table.add_column("Actor",key="actor")
+        table.add_column("Status",key="status")
+        table.add_column("pid",key="pid")
+        table.add_column("CPU",key="cpu")
+        table.add_column("Mem",key="mem")
+        table.add_column("Started",key="started")
+        
         self._refresh_timer = self.set_interval(2, self.update_status)
 
+
+        statuses = ActorSystemConnector.get_status()
+        for status_item in statuses:
+            table.add_row(
+                        status_item.actor_name,
+                        status_item.status,
+                        status_item.pid,
+                        status_item.cpu_usage,
+                        status_item.human_readable_memory(),
+                        status_item.started,
+                        key=str(status_item.pid),
+                    )
         # self.name = configuration["name"]
         # self.id = configuration["id"]
         # self.run_number = run_number
@@ -117,25 +135,51 @@ class MTreeSystemStatusScreen(Screen):
         # self.end_time = None
 
     def update_status(self) -> None:
-        actor_system = ActorSystemConnector()
-        statuses = actor_system.get_status()
-
-        status_table = self.query_one("#system-status-table")
-
-        for run_status in statuses:
+        
+        table = self.query_one(DataTable)
+        statuses = ActorSystemConnector.get_status()
+        for status_item in statuses:
             try:
-                row_index = status_table.get_row_index(run_status[0])
-                for index, i in enumerate(run_status):
-                    status_table.update_cell_at(Coordinate(row_index, index), i)
+                table.get_row(str(status_item.pid))
+                table.update_cell(str(status_item.pid), "status", status_item.status)
+                table.update_cell(str(status_item.pid), "cpu", status_item.cpu_usage)
+                table.update_cell(str(status_item.pid), "mem", status_item.human_readable_memory())
+
+                # table.update_cell_at(Coordinate(row_index, 0), status_item.actor_name)
+                # table.update_cell_at(Coordinate(row_index, 1), status_item.status)
+                # table.update_cell_at(Coordinate(row_index, 2), status_item.pid)
+                # table.update_cell_at(Coordinate(row_index, 3), status_item.cpu_usage)
+                # table.update_cell_at(Coordinate(row_index, 4), status_item.memory_usage)
+                # table.update_cell_at(Coordinate(row_index, 5), status_item.started)
             except:
-                status_table.add_row(
-                    run_status[0],
-                    run_status[1],
-                    run_status[2],
-                    run_status[3],
-                    run_status[4],
-                    key=run_status[0],
-                )
+                table.add_row(
+                        status_item.actor_name,
+                        status_item.status,
+                        status_item.pid,
+                        status_item.cpu_usage,
+                        status_item.human_readable_memory(),
+                        status_item.started,
+                        key=str(status_item.pid),
+                    )
+
+            
+        
+
+
+        # for run_status in statuses:
+        #     try:
+        #         row_index = status_table.get_row_index(run_status[0])
+        #         for index, i in enumerate(run_status):
+        #             status_table.update_cell_at(Coordinate(row_index, index), i)
+        #     except:
+        #         status_table.add_row(
+        #             run_status[0],
+        #             run_status[1],
+        #             run_status[2],
+        #             run_status[3],
+        #             run_status[4],
+        #             key=run_status[0],
+        #         )
 
         last_status_check = self.query_one("#last-status-check")
         last_status_check.update(f"System Last Checked: {datetime.now().isoformat()}")
