+++ pyproject.toml
@@ -28,9 +28,10 @@ dependencies = [
     'dnspython',
     'bump-my-version==0.19.0',
     'towncrier==23.11.0',
-    'setproctitle==1.3.3'
+    'setproctitle==1.3.3',
+    'psutil==5.9.8'
 ]
-requires-python = ">=3.11"
+requires-python = ">=3.10"
 authors = [
   {name = "GMU CEELAB", email = "mtree.email@gmail.com"}
 ]
